Update to v0r1 release.
byuu says:

So I tried to make a new release of beat, but unfortunately when trying to make
a large patch (300MB ISO to 320MB ISO), the created beat patches are invalid.
Applying the patch corrupts about ~1% of the file, because the suffix array
isn't sorted correctly. But it only happens in the extreme case with huge files,
seemingly.